move parallel-installation files to separate dir
