+++ components/provisioner/internal/parallel-installation/components_status.go
@@ -1,4 +1,4 @@
-package installation
+package parallel_installation
 
 import (
 	"fmt"
