+++ components/provisioner/internal/parallel-installation/components_status_test.go
@@ -1,4 +1,4 @@
-package installation
+package parallel_installation
 
 import (
 	"testing"
