+++ components/provisioner/internal/parallel-installation/download/components/testdata/simple-chart.tgz
