+++ components/provisioner/internal/parallel-installation/download/components/zip.go
