+++ components/provisioner/internal/parallel-installation/mocks/ResourceDownloader.go
