+++ components/provisioner/internal/parallel-installation/overrides.go
@@ -1,4 +1,4 @@
-package installation
+package parallel_installation
 
 import (
 	"fmt"
