+++ components/provisioner/internal/parallel-installation/overrides_test.go
@@ -1,10 +1,10 @@
-package installation
+package parallel_installation
 
 import (
 	"testing"
 
-	"github.com/kyma-project/control-plane/components/provisioner/internal/installation/mocks"
 	"github.com/kyma-project/control-plane/components/provisioner/internal/model"
+	"github.com/kyma-project/control-plane/components/provisioner/internal/parallel-installation/mocks"
 
 	"github.com/stretchr/testify/require"
 )
